+++ lib/models/user_update_model.dart
@@ -1,4 +1,5 @@
 
+import 'package:runetid_sdk/models/event_participant.dart';
 import 'package:runetid_sdk/models/user.dart';
 
 class UserUpdateModel {